Moved common function to new invitation utils file
